requirements.txt
streamlit
transformers
torch
pillow
langchain
langchain-google-genai
python-dotenv
